requirements.txt
mcp>=1.26.0
pywin32>=311
watchdog>=6.0.0
pyyaml>=6.0
requests>=2.31
